[Style] 내모여락 관련 네이밍 수정- MyGatherock, Gatherock, event 관련 변수명 Gathering으로 수정 (#30)
