+++ GetARock/GetARock/Global/Resource/Xibs/GatheringCell.xib
@@ -4,12 +4,13 @@
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="20020"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="GatherockListCell" rowHeight="118" id="S08-fi-dea" customClass="GatherockListCell" customModule="GetARock" customModuleProvider="target">
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="GatheringCell" rowHeight="118" id="S08-fi-dea" customClass="GatheringCell" customModule="GetARock" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="414" height="100"/>
             <autoresizingMask key="autoresizingMask"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="S08-fi-dea" id="0NC-3l-Vgg">
@@ -70,11 +71,11 @@
                 </constraints>
             </tableViewCellContentView>
             <connections>
-                <outlet property="eventState" destination="Toa-R0-wqa" id="okB-9p-OoT"/>
-                <outlet property="eventTitle" destination="4Wl-Jv-4Rb" id="I8E-pr-Jvo"/>
-                <outlet property="startTime" destination="QDG-Zm-Xit" id="wmc-L2-G70"/>
+                <outlet property="startTime" destination="QDG-Zm-Xit" id="IWd-Ev-Hsv"/>
+                <outlet property="state" destination="Toa-R0-wqa" id="0dX-vd-g5e"/>
+                <outlet property="title" destination="4Wl-Jv-4Rb" id="syM-Eg-yqB"/>
             </connections>
-            <point key="canvasLocation" x="137.68115942028987" y="113.16964285714285"/>
+            <point key="canvasLocation" x="443" y="125"/>
         </tableViewCell>
     </objects>
     <resources>
