+++ GetARock/GetARock/Screen/Gathering/GatheringCreatedViewController.swift
@@ -1,0 +1,79 @@
+//
+//  GatherockListCreatedViewController.swift
+//  GetARock
+//
+//  Created by Admin on 2022/11/18.
+//
+
+import UIKit
+
+// MARK: - sample data
+enum stateType {
+    case finding
+    case active
+    case finish
+    case cancel
+}
+
+struct Gathering {
+    var title: String
+    var date: String
+    var state: stateType
+    
+    init(title: String, date: String, state: stateType){
+        self.title = title
+        self.date = date
+        self.state = state
+    }
+}
+
+class GatheringCreatedViewController: UIViewController {
+    @IBOutlet weak var tableView: UITableView!
+    
+    private var gathering = [
+        Gathering(title: "동성로 합동 버스킹 합니다!", date: "22.12.05 13:00", state: .finding),
+        Gathering(title: "블랙로즈 합주 드럼 한 분 구합니다", date: "22.11.20 14:00", state: .active),
+        Gathering(title: "몰라 아무거나 쳐보는 중입니다.", date: "22.11.20 12:00", state: .active),
+        Gathering(title: "두 줄 만들기 해보는 중, 몰라 아무거나 쳐보는 중입니다1234.", date: "22.11.18 21:00", state: .cancel),
+        Gathering(title: "대체텍스트 대체텍스트 대체텍스트", date: "22.11.18 16:00", state: .finish),
+        Gathering(title: "두 줄 만들기 대체텍스트 대체텍스트 대체텍스트 대체텍스트 대체텍스트 22", date: "22.11.12 18:00", state: .finish),
+    ]
+    
+    func buttonType(for type: stateType) -> String {
+        let text: String
+        switch type {
+        case .finding: text = "모집중"
+        case .active: text = "진행중"
+        case .finish: text = "완 료"
+        case .cancel: text = "취소됨"
+        }
+        return text
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        let nibName = UINib(nibName: "GatheringCell", bundle: nil)
+        tableView.register(nibName, forCellReuseIdentifier: GatheringCell.className)
+    }
+    
+}
+
+extension GatheringCreatedViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return gathering.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: GatheringCell.className, for: indexPath) as! GatheringCell
+        
+        cell.title.text = gathering[indexPath.row].title
+        cell.startTime.text = gathering[indexPath.row].date
+        cell.state.setTitle(buttonType(for: gathering[indexPath.row].state), for: .normal)
+        return cell
+    }
+}
+
+extension GatheringCreatedViewController: UITableViewDelegate {
+    
+}
